feat(backend): added more request to the api regarding post-logic and started a draft on rating
Added more functions to postHandler.ts and created a ratingHandler.ts. Changed the backend server
port to 5001. Updated string formatting on the seed. Also changed formatting for end of line
sequence from CRLF to LF in order to satisfy esLint

BREAKING CHANGE: The code is a draft and is not working

Related to #8 #11