feat(frontend): wrote a Test component for testing GET request from frontend
Test.tsx has been added in order to test simple fetching from localhost (GET request). Had to add
cors to yarn in order for this to work, therefore yarn.lock and package.json are affected. Test can
be used to get the info of a specified userID by a click of a button.

Related to #21
Development-method: individual programming