feat((frontend)): created logic for sending PostCreationData from form through middleware to backend
Changed the form in CreatePosts.tsx to use ToggleButtons and added a Datepicker, City and Venue. The
image input has temporarily been removed. When submitting the form the user now gets either an Alert
saying something is wrong witht the input, or the user gets sent to the profile page. There is not
much validation added and there is no code for cleaning up data in postHandler. Axios and React
Datepicker have been added to yarn.

Closes #32
Related to #29