feat(frontend): Changed post, name shows on the post, hardcoded profile page to other users.
Did some changes to post, so the name can be shown on the top of the post. Created UserPage that should send you to user page by clicking on the name on top of the post. It is hardcoded and sends you to user 500 so far, see App.tsx and UserPage.tsx.